requirements.txt
flask>=3.0
deep-translator>=1.11
easyocr>=1.7
PyMuPDF>=1.24
Pillow>=10.0
numpy>=1.24
